refactor: move static_file_service into files namespace
