+++ tests/static_file_service_test.php
@@ -34,6 +34,7 @@ use invalid_dataroot_permissions;
 use moodle_exception;
 use qtype_questionpy\local\api\api;
 use qtype_questionpy\local\api\qpy_http_client;
+use qtype_questionpy\local\files\static_file_service;
 
 /**
  * Tests QuestionPy static file access.
@@ -47,7 +48,7 @@ use qtype_questionpy\local\api\qpy_http_client;
  * @copyright  2024 TU Berlin, innoCampus {@link https://www.questionpy.org}
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  *
- * @covers \qtype_questionpy\static_file_service
+ * @covers \qtype_questionpy\local\files\static_file_service
  * @covers \qtype_questionpy\local\api\package_api::download_static_file
  */
 final class static_file_service_test extends \advanced_testcase {
